requirements.txt
aiohttp==3.5.4
astropy==3.1.2
async-timeout==3.0.1
attrs==19.1.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
docutils==0.14
ecdsa==0.13.2
envs==1.3
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
jmespath==0.9.4
multidict==4.5.2
numpy==1.16.4
pycparser==2.19
pycryptodome==3.3.1
pypiwin32==223
python-dateutil==2.8.0
python-dotenv==0.10.2
python-jose-cryptodome==1.3.2
pywin32==224
requests==2.22.0
rope==0.14.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.25.3
warrant==0.6.1
yarl==1.3.0
